requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
python-dotenv
requests
psycopg2-binary